builds the div for a reactive shiny value box

builds the div for a reactive shiny value box

Usage

vbox_reactive(
  label = "label",
  textid = "textOutput_here",
  col = "rgba(60,130,180,0.5)"
)

Arguments

label

The value box header label to appear- character of length one

textid

The value - A character vector of length one, optionally the textOutput of a reactive value if using vbox_reactive

col

Bkgrd col- A character of rbg(), rgba(), or a chtml olor name to recognixed in R

Value

A character vector of the div tree to be rendered in the ui of a shiny app

Examples

vbox_reactive('Label','YourTextOutputIdHere')
#> <div style="border-radius: 15px;&#10;                      border-style: solid;&#10;                      border-color:white;&#10;            padding:0px;&#10;            margin:0px 10px;&#10;            width:200px;&#10;            color:white;&#10;                  font-weight: 300;&#10;                  text-color:blue;&#10;                  text-align:center;&#10;                  background-color:rgba(60,130,180,0.5);">
#>   <h5 style="color:white;">Label</h5>
#>   <h2 style="font-weight:bold;display: inline-block;vertical-align:top;color:white;">
#>     <div id="YourTextOutputIdHere" class="shiny-text-output"></div>
#>   </h2>
#> </div>